refactor auditor
Move it to the `db` package (it was in the `payments` package for
historical reasons but doesn't deal only with payment anymore).

Better typing for channel lifecycle event.
