+++ eclair-core/src/main/scala/fr/acinq/eclair/db/AuditDb.scala
@@ -16,12 +16,14 @@
 
 package fr.acinq.eclair.db
 
-import java.io.Closeable
-
 import fr.acinq.bitcoin.Crypto.PublicKey
 import fr.acinq.bitcoin.{ByteVector32, Satoshi}
+import fr.acinq.eclair.channel.Helpers.Closing._
 import fr.acinq.eclair.channel._
+import fr.acinq.eclair.db.AuditDb.{ChannelLifecycleEvent, NetworkFee, Stats}
 import fr.acinq.eclair.payment.{PaymentReceived, PaymentRelayed, PaymentSent}
+
+import java.io.Closeable
 
 trait AuditDb extends Closeable {
 
@@ -49,8 +51,31 @@ trait AuditDb extends Closeable {
 
 }
 
-case class ChannelLifecycleEvent(channelId: ByteVector32, remoteNodeId: PublicKey, capacity: Satoshi, isFunder: Boolean, isPrivate: Boolean, event: String)
+object AuditDb {
 
-case class NetworkFee(remoteNodeId: PublicKey, channelId: ByteVector32, txId: ByteVector32, fee: Satoshi, txType: String, timestamp: Long)
+  // @formatter:off
+  case class ChannelLifecycleEvent(channelId: ByteVector32, remoteNodeId: PublicKey, capacity: Satoshi, isFunder: Boolean, isPrivate: Boolean, event: ChannelLifecycleEvent.EventType)
+  object ChannelLifecycleEvent {
+    sealed trait EventType { def label: String }
+    object EventType {
+      object Created extends EventType { override def label: String = "created" }
+      object Connected extends EventType { override def label: String = "connected" }
+      case class Closed(closingType: ClosingType) extends EventType {
+        override def label: String = closingType match {
+          case _: MutualClose => "mutual"
+          case _: LocalClose => "local"
+          case _: CurrentRemoteClose => "remote"
+          case _: NextRemoteClose => "remote"
+          case _: RecoveryClose => "recovery"
+          case _: RevokedClose => "revoked"
+        }
+      }
+    }
+  }
+  // @formatter:on
 
-case class Stats(channelId: ByteVector32, direction: String, avgPaymentAmount: Satoshi, paymentCount: Int, relayFee: Satoshi, networkFee: Satoshi)
+  case class NetworkFee(remoteNodeId: PublicKey, channelId: ByteVector32, txId: ByteVector32, fee: Satoshi, txType: String, timestamp: Long)
+
+  case class Stats(channelId: ByteVector32, direction: String, avgPaymentAmount: Satoshi, paymentCount: Int, relayFee: Satoshi, networkFee: Satoshi)
+
+}
