+++ eclair-core/src/main/scala/fr/acinq/eclair/db/Auditor.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 ACINQ SAS
+ * Copyright 2021 ACINQ SAS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package fr.acinq.eclair.payment
+package fr.acinq.eclair.db
 
 import akka.actor.{Actor, ActorLogging, Props}
 import fr.acinq.eclair.NodeParams
-import fr.acinq.eclair.channel.Helpers.Closing._
 import fr.acinq.eclair.channel.Monitoring.{Metrics => ChannelMetrics, Tags => ChannelTags}
 import fr.acinq.eclair.channel._
-import fr.acinq.eclair.db.ChannelLifecycleEvent
+import fr.acinq.eclair.db.AuditDb.ChannelLifecycleEvent
 import fr.acinq.eclair.payment.Monitoring.{Metrics => PaymentMetrics, Tags => PaymentTags}
+import fr.acinq.eclair.payment._
 
 class Auditor(nodeParams: NodeParams) extends Actor with ActorLogging {
 
@@ -82,7 +82,8 @@ class Auditor(nodeParams: NodeParams) extends Actor with ActorLogging {
       e match {
         case ChannelStateChanged(_, channelId, _, remoteNodeId, WAIT_FOR_FUNDING_LOCKED, NORMAL, Some(commitments: Commitments)) =>
           ChannelMetrics.ChannelLifecycleEvents.withTag(ChannelTags.Event, ChannelTags.Events.Created).increment()
-          db.add(ChannelLifecycleEvent(channelId, remoteNodeId, commitments.capacity, commitments.localParams.isFunder, !commitments.announceChannel, "created"))
+          val event = ChannelLifecycleEvent.EventType.Created
+          db.add(ChannelLifecycleEvent(channelId, remoteNodeId, commitments.capacity, commitments.localParams.isFunder, !commitments.announceChannel, event))
         case ChannelStateChanged(_, _, _, _, WAIT_FOR_INIT_INTERNAL, _, _) =>
         case ChannelStateChanged(_, _, _, _, _, CLOSING, _) =>
           ChannelMetrics.ChannelLifecycleEvents.withTag(ChannelTags.Event, ChannelTags.Events.Closing).increment()
@@ -91,13 +92,7 @@ class Auditor(nodeParams: NodeParams) extends Actor with ActorLogging {
 
     case e: ChannelClosed =>
       ChannelMetrics.ChannelLifecycleEvents.withTag(ChannelTags.Event, ChannelTags.Events.Closed).increment()
-      val event = e.closingType match {
-        case _: MutualClose => "mutual"
-        case _: LocalClose => "local"
-        case _: RemoteClose => "remote" // can be current or next
-        case _: RecoveryClose => "recovery"
-        case _: RevokedClose => "revoked"
-      }
+      val event = ChannelLifecycleEvent.EventType.Closed(e.closingType)
       db.add(ChannelLifecycleEvent(e.channelId, e.commitments.remoteParams.nodeId, e.commitments.commitInput.txOut.amount, e.commitments.localParams.isFunder, !e.commitments.announceChannel, event))
 
   }
@@ -108,6 +103,6 @@ class Auditor(nodeParams: NodeParams) extends Actor with ActorLogging {
 
 object Auditor {
 
-  def props(nodeParams: NodeParams) = Props(classOf[Auditor], nodeParams)
+  def props(nodeParams: NodeParams): Props = Props(new Auditor(nodeParams))
 
 }
